requirements.txt
Flask~=3.1.0
requests~=2.32.3
PyMuPDF~=1.24.2
Werkzeug~=3.1.3
us~=3.2.0
geopy~=2.4.1
flask-cors~=6.0.0
python-json-logger==3.3.0
PyYAML==6.0.2
setuptools
wheel
